vhost: fix compilation issue when vhost debug enabled
This patch fixes below compilation issue when
RTE_LIBRTE_VHOST_DEBUG is enabled:

In file included from ./lib/librte_vhost/virtio_net.c:50:0:
./lib/librte_vhost/virtio_net.c: In function ‘copy_mbuf_to_desc’:
./lib/librte_vhost/vhost.h:205:42: warning: format ‘%d’ expects argument of type ‘int’, but argument 5 has type ‘uint64_t {aka long unsigned int}’ [-Wformat=]
   snprintf(packet, VHOST_MAX_PRINT_BUFF, "(%d) Header size %d: ", (device->vid), (size)); \
                                          ^
./lib/librte_vhost/virtio_net.c:301:4: note: in expansion of macro ‘PRINT_PACKET’
    PRINT_PACKET(dev, (uintptr_t)dst, len, 0);
    ^~~~~~~~~~~~

Fixes: 067047a90055 ("vhost: handle virtually non-contiguous buffers in Tx")
Fixes: dd906b814d9a ("vhost: handle virtually non-contiguous buffers in Rx")
Fixes: 2b8d39414619 ("vhost: handle virtually non-contiguous buffers in Rx-mrg")

Acked-by: Luca Boccassi <bluca@debian.org>

+++ lib/librte_vhost/virtio_net.c
@@ -298,7 +298,7 @@ copy_mbuf_to_desc(struct virtio_net *dev, struct vring_desc *descs,
 			rte_memcpy((void *)(uintptr_t)dst,
 					(void *)(uintptr_t)src, len);
 
-			PRINT_PACKET(dev, (uintptr_t)dst, len, 0);
+			PRINT_PACKET(dev, (uintptr_t)dst, (uint32_t)len, 0);
 			remain -= len;
 			guest_addr += len;
 			dst += len;
@@ -736,7 +736,7 @@ copy_mbuf_to_desc_mergeable(struct virtio_net *dev, struct rte_mbuf *m,
 							len);
 
 					PRINT_PACKET(dev, (uintptr_t)dst,
-							len, 0);
+							(uint32_t)len, 0);
 
 					remain -= len;
 					guest_addr += len;
@@ -1116,7 +1116,7 @@ copy_desc_to_mbuf(struct virtio_net *dev, struct vring_desc *descs,
 	rte_prefetch0((void *)(uintptr_t)(desc_addr + desc_offset));
 
 	PRINT_PACKET(dev, (uintptr_t)(desc_addr + desc_offset),
-			desc_chunck_len, 0);
+			(uint32_t)desc_chunck_len, 0);
 
 	mbuf_offset = 0;
 	mbuf_avail  = m->buf_len - RTE_PKTMBUF_HEADROOM;
@@ -1181,7 +1181,7 @@ copy_desc_to_mbuf(struct virtio_net *dev, struct vring_desc *descs,
 			desc_avail  = desc->len;
 
 			PRINT_PACKET(dev, (uintptr_t)desc_addr,
-					desc_chunck_len, 0);
+					(uint32_t)desc_chunck_len, 0);
 		} else if (unlikely(desc_chunck_len == 0)) {
 			desc_chunck_len = desc_avail;
 			desc_gaddr += desc_offset;
@@ -1194,7 +1194,7 @@ copy_desc_to_mbuf(struct virtio_net *dev, struct vring_desc *descs,
 			desc_offset = 0;
 
 			PRINT_PACKET(dev, (uintptr_t)desc_addr,
-					desc_chunck_len, 0);
+					(uint32_t)desc_chunck_len, 0);
 		}
 
 		/*
